requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.1
future==0.18.2
objects==0.3.1
pkg_resources==0.0.0
PyJWT==1.7.1
pytz==2022.5
router==0.1
serializers==0.2.4
sqlparse==0.4.3
